calSlope default to 950
My testing found that the lower calSlope (hard-coded to 750 or 850) led to more aggressive behavior at higher glucose levels. To be safer, I'm setting this to 950 to lessen that default behavior. Users can modify it directly for now if necessary at least until we enhance logger to use xdrip+ level calibration.