native-calibrates-lsr mode
adds native-calibrates-lsr mode, which calibrates the LSR algo. every 6 hours on the 6 hours, from a native TX reading (g5 or g6).
also skips sending any calibrations to a TX if sensorCode is set in config, for g6 no-calibration mode.
(note: if you find your no-calibration mode is not reliable you'll have to remove sensorCode from the config to calibrate it from logger).