Remove false disclaimer
The claim that Logger is not intended for therapy is untrue.  As evidenced by all of the other documentation, and all of the public conversations around its development, it *is* intended for use with OpenAPS, as part of a DIY artificial insulin delivery system self-built by the user.  Neither Logger nor OpenAPS is distributed as a product, but rather as source code protected under the first amendment to the US constitution.

If you want a strong disclaimer, you could point out that Logger is not a product, comes with no warranty or official support, that anyone using it is doing so at their own risk and must take responsibility for their own safety, and that use of Logger for therapy is not FDA approved and comes with inherent risks.